Implement exporting to json asynchronously

+++ src/export/export.h
@@ -1,11 +1,11 @@
 #ifndef EXPORT_H
 #define EXPORT_H
 
-#include <stddef.h>
+#include <stdint.h>
 
 #include "queue.h"
 
-int export_start_thread(Queue* queue, int export_rate_seconds, size_t export_treshold);
+int export_start_thread(Queue* queue, int export_rate_seconds, uint32_t export_treshold);
 int export_stop_thread();
 
 #endif
